requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
fastapi==0.110.0
gotrue==2.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
packaging==24.0
pillow==10.2.0
postgrest==0.16.2
pydantic==2.6.4
pydantic_core==2.16.3
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
qrcode==7.4.2
realtime==1.0.2
redis==5.0.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.0
supafunc==0.3.3
typing_extensions==4.10.0
uvicorn==0.19.0
websockets==11.0.3
